Avoid KPD server stop by PR update failure
It is often observed that KPD stops unexpectedly after the INFO message
like below:

   "Waiting for PullRequest(X) sha=Y to go to Z"

The message is printed in BranchWorker.apply_push_comment(). It is not
yet understood why GitHub fails to update the PR after force push.

To avoid the KPD stop, do not throw RuntimeError. Instead, close the PR
to ignore.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>